1_Trainee/src/main/java/collections/framework/search/PhoneBook.java
package collections.framework.search;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

/**
 * Phone book.
 * @author Egor Novikov
 * E-mail: enovikovdev@gmail.com
 * @version 1
 * @since 1
 */
public class PhoneBook {
    /**
     * Phone book.
     */
    private List<Person> book = new ArrayList<Person>();
    /**
     * Method adds person to phone directory.
     * @param person - person.
     */
    public void add(Person person) {
        this.book.add(person);
    }
    /**
     * Find person by key.
     * @param key - key.
     * @return - person.
     */
    public List<Person> find(String key) {
        List<Person> result = book.stream().filter(person -> person.getName().contains(key)
                || person.getSurname().contains(key)
                || person.getPhone().contains(key)
                || person.getAddress().contains(key)
        ).collect(Collectors.toList());
        return result;
    }
}